requirement.txt
appdirs==1.4.4
audiomentations==0.30.0
audioread==3.0.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
ffmpeg-python==0.2.0
fonttools==4.38.0
future==0.18.3
idna==3.4
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
librosa==0.9.2
llvmlite==0.39.1
matplotlib==3.6.3
mypy-extensions==0.4.3
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
pooch==1.6.0
primePy==1.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
resampy==0.4.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
soundfile==0.11.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.2
torchaudio==0.13.1
torchsummary==1.5.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
